+++ src/network/network_gui.cpp
@@ -1179,7 +1179,7 @@ struct NetworkStartServerWindow : public Window {
 			}
 
 			case WID_NSS_GENERATE_GAME: // Start game
-				if ((uint) CountSelectedGRFs (_grfconfig_newgame) > NETWORK_MAX_GRF_COUNT) {
+				if (CountSelectedGRFs(_grfconfig_newgame) > NETWORK_MAX_GRF_COUNT) {
 					ShowErrorMessage(STR_NEWGRF_ERROR_TOO_MANY_NEWGRFS_LOADED, INVALID_STRING_ID, WL_ERROR);
 					break;
 				}
